# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Operator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/sessions:
    get:
      operationId: listSessions
      parameters:
      - description: Project id filter.
        in: query
        name: project
        schema:
          description: Project id filter.
          type: string
      - description: When true, return non-terminated sessions; when false, return
          terminated sessions.
        in: query
        name: active
        schema:
          description: When true, return non-terminated sessions; when false, return
            terminated sessions.
          type:
          - "null"
          - boolean
      - description: When true, return only orchestrator sessions.
        in: query
        name: orchestratorOnly
        schema:
          description: When true, return only orchestrator sessions.
          type:
          - "null"
          - boolean
      - description: When true, return only fresh non-terminated sessions.
        in: query
        name: fresh
        schema:
          description: When true, return only fresh non-terminated sessions.
          type:
          - "null"
          - boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List sessions
      tags:
      - sessions
    post:
      operationId: spawnSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpawnSessionRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpawnSessionResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Spawn a new agent session
      tags:
      - sessions
  /api/v1/sessions/{sessionId}:
    get:
      operationId: getSession
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Fetch one session
      tags:
      - sessions
    patch:
      operationId: renameSession
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameSessionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenameSessionResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Rename a session display name
      tags:
      - sessions
...
  /api/v1/settings:
    get:
      operationId: getSettings
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Read the daemon-owned user preferences
      tags:
      - settings
  /api/v1/settings/session-interface:
    patch:
      operationId: updateSessionInterface
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionInterfaceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Choose the default interface for new sessions
      tags:
      - settings
...
  /api/v1/usage/sessions:
    get:
      operationId: listCompactSessionUsage
      parameters:
      - description: Optional project id filter for dashboard cards.
        in: query
        name: projectId
        schema:
          description: Optional project id filter for dashboard cards.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCompactSessionUsageResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List compact token usage for session cards
      tags:
      - usage
  /api/v1/usage/sessions/{sessionId}:
    get:
      operationId: getSessionUsage
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUsageResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get detailed token usage for one session
      tags:
      - usage
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AttachmentInput:
      properties:
        data:
          type: string
        mimeType:
          type: string
      required:
      - data
      type: object
...
    CompactSessionUsageResponse:
      properties:
        incomplete:
          type: boolean
        sessionId:
          type: string
        totalTokens:
          format: int64
          minimum: 0
          type: integer
      required:
      - sessionId
      - totalTokens
      - incomplete
      type: object
...
    ControllersSessionView:
      properties:
        activity:
          $ref: '#/components/schemas/DomainActivity'
        autoInjectReview:
          type: boolean
        branch:
          type: string
        createdAt:
          format: date-time
          type: string
        displayName:
          type: string
        harness:
          type: string
        id:
          type: string
        isPinned:
          type: boolean
        isTerminated:
          type: boolean
        issueId:
          type: string
        kind:
          type: string
        mode:
          enum:
          - chat
          - tui
          type: string
        pinnedAt:
          format: date-time
          type:
          - "null"
          - string
        previewRevision:
          format: int64
          type: integer
        previewUrl:
          type: string
        projectId:
          type: string
        prs:
          items:
            $ref: '#/components/schemas/SessionPRFacts'
          type: array
        reviewerHarness:
          enum:
          - claude-code
          - codex
          - copilot
          - cursor
          - kilocode
          - opencode
          - kiro
          - pi
          - qwen
          - agy
          - continue
          - goose
          - vibe
          - devin
          - droid
          - kimi
          - kimchi
          - muse
          - amp
          - aider
          - grok
          - crush
          - auggie
          - cline
          - autohand
          type: string
        scmStatus:
          enum:
          - pr_open
          - draft
          - ci_failed
          - review_pending
          - changes_requested
          - approved
          - mergeable
          - merged
          type: string
        status:
          enum:
          - working
          - pr_open
          - draft
          - ci_failed
          - review_pending
          - changes_requested
          - approved
          - mergeable
          - merged
          - needs_input
          - exited
          - idle
          - terminated
          - no_signal
          type: string
        terminalHandleId:
          type: string
        terminateOnPrMerge:
          type: boolean
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - projectId
      - kind
      - mode
      - activity
      - isTerminated
      - terminateOnPrMerge
      - autoInjectReview
      - createdAt
      - updatedAt
      - isPinned
      - status
      - prs
      type: object
...
    DomainActivity:
      properties:
        lastActivityAt:
          format: date-time
          type: string
        state:
          type: string
      required:
      - state
      - lastActivityAt
      type: object
...
    ListCompactSessionUsageResponse:
      properties:
        sessions:
          items:
            $ref: '#/components/schemas/CompactSessionUsageResponse'
          type: array
      required:
      - sessions
      type: object
...
    ListSessionsResponse:
      properties:
        sessions:
          items:
            $ref: '#/components/schemas/ControllersSessionView'
          type: array
      required:
      - sessions
      type: object
...
    RenameSessionRequest:
      properties:
        displayName:
          minLength: 1
          type: string
      required:
      - displayName
      type: object
    RenameSessionResponse:
      properties:
        displayName:
          type: string
        ok:
          type: boolean
        sessionId:
          type: string
      required:
      - ok
      - sessionId
      - displayName
      type: object
...
    SessionPRFacts:
      properties:
        ci:
          enum:
          - unknown
          - pending
          - passing
          - failing
          type: string
        mergeability:
          enum:
          - unknown
          - mergeable
          - conflicting
          - blocked
          - unstable
          type: string
        number:
          type: integer
        review:
          enum:
          - none
          - approved
          - changes_requested
          - review_required
          type: string
        reviewComments:
          type: boolean
        state:
          enum:
          - draft
          - open
          - merged
          - closed
          type: string
        updatedAt:
          format: date-time
          type: string
        url:
          type: string
      required:
      - url
      - number
      - state
      - ci
      - review
      - mergeability
      - reviewComments
      - updatedAt
      type: object
...
    SessionResponse:
      properties:
        session:
          $ref: '#/components/schemas/ControllersSessionView'
      required:
      - session
      type: object
    SessionUsageResponse:
      properties:
        harnesses:
          items:
            $ref: '#/components/schemas/UsageHarnessResponse'
          type: array
        incomplete:
          type: boolean
        sessionId:
          type: string
        totals:
          $ref: '#/components/schemas/UsageTotalsResponse'
      required:
      - sessionId
      - incomplete
      - totals
      - harnesses
      type: object
...
    SettingsResponse:
      properties:
        chatHarnesses:
          items:
            type: string
          type: array
        defaultSessionMode:
          enum:
          - chat
          - tui
          type: string
      required:
      - defaultSessionMode
      - chatHarnesses
      type: object
...
    SpawnSessionRequest:
      properties:
        attachments:
          items:
            $ref: '#/components/schemas/AttachmentInput'
          type: array
        branch:
          type: string
        displayName:
          maxLength: 20
          type: string
        harness:
          enum:
          - claude-code
          - codex
          - aider
          - opencode
          - grok
          - droid
          - amp
          - agy
          - crush
          - cursor
          - qwen
          - copilot
          - goose
          - auggie
          - continue
          - devin
          - cline
          - kimi
          - muse
          - kiro
          - kilocode
          - vibe
          - pi
          - kimchi
          - prime-agent
          - autohand
          type: string
        issueId:
          type: string
        kind:
          enum:
          - worker
          - orchestrator
          type: string
        mode:
          enum:
          - chat
          - tui
          type: string
        projectId:
          type: string
        prompt:
          maxLength: 4096
          type: string
      required:
      - projectId
      type: object
    SpawnSessionResponse:
      properties:
        promptBytes:
          type: integer
        session:
          $ref: '#/components/schemas/ControllersSessionView'
        systemPromptBytes:
          type: integer
      required:
      - session
      - promptBytes
      - systemPromptBytes
      type: object
...
    UpdateSessionInterfaceRequest:
      properties:
        defaultSessionMode:
          enum:
          - chat
          - tui
          type: string
      required:
      - defaultSessionMode
      type: object
...
    UsageHarnessResponse:
      properties:
        harness:
          type: string
        models:
          items:
            $ref: '#/components/schemas/UsageModelResponse'
          type: array
        totals:
          $ref: '#/components/schemas/UsageTotalsResponse'
      required:
      - harness
      - totals
      - models
      type: object
...
    UsageModelResponse:
      properties:
        modelId:
          type: string
        totals:
          $ref: '#/components/schemas/UsageTotalsResponse'
      required:
      - modelId
      - totals
      type: object
    UsageTotalsResponse:
      properties:
        cacheReadTokens:
          type:
          - "null"
          - integer
        cacheWriteTokens:
          type:
          - "null"
          - integer
        inputTokens:
          type:
          - "null"
          - integer
        outputTokens:
          type:
          - "null"
          - integer
        reasoningTokens:
          type:
          - "null"
          - integer
        uncachedInputTokens:
          type:
          - "null"
          - integer
      required:
      - inputTokens
      - uncachedInputTokens
      - cacheReadTokens
      - cacheWriteTokens
      - outputTokens
      - reasoningTokens
      type: object
...